% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdbGetShow.R
\name{cdbGetShow}
\alias{cdbGetShow}
\title{Receive show results from CouchDB}
\usage{
cdbGetShow(cdb)
}
\arguments{
\item{cdb}{Beside the connection details (\code{cdb$port}, \code{cdb$DBName}
...) the \code{cdb$design} and \code{cdb$show} is needed.}
}
\value{
\item{cdb }{The result of the request is stored in cdb$res after converting
the json answer into a list using \code{cdb$fromJSON()}. If a needed cdb
entry was not provided \code{cdb$error} provides information.

}
}
\description{
The function provides accesses to CouchDB shows.
}
\details{
Query params e.g. \code{"format=json"}  can be
provided in \code{cdb$queryParam}. Multible params must be given in
one string e.g.  \code{"a=b&c=d&e=f"}.
}
\author{
wactbprot
}
\keyword{misc}

