% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdbDeleteDoc.R
\name{cdbDeleteDoc}
\alias{cdbDeleteDoc}
\title{Deletes a document from a database}
\usage{
cdbDeleteDoc(cdb)
}
\arguments{
\item{cdb}{Beside \code{cdb$serverName}, \code{cdb$port} and
\code{cdb$DBName} the \code{cdb$id} must be given. R errors are reported in
\code{cdb$errors}}
}
\value{
\item{cdb }{The result of the delete request is stored in
\code{cdb$res}(whatever this means).  }
}
\description{
With a given \code{cdb$id} this function sends a http \code{"DELETE"}
request to the url \code{.../cdb$id?rev=cdb$rev}.
}
\author{
wactbprot
}
\seealso{
\code{cdbAddDoc()}
}
\keyword{misc}

