% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cdbGetUuidS}
\alias{cdbGetUuidS}
\title{Function for request some ids}
\usage{
cdbGetUuidS(cdb)
}
\arguments{
\item{cdb}{Only the connection settings \code{cdb$port},
\code{cdb$serverName} and \code{cdb$count} is needed.}
}
\value{
\item{cdb }{The result of the request is stored in \code{cdb$res} after
converting the answer into a list using \code{fromJSON()}.}
}
\description{
Function returns a 128bit uuid requested from CouchDB
}
\details{
CouchDB API provides the url http://serverName:port/_uuids for those clients
who aren't able to create those ids. The number N
of ids received from a CouchDB can be set by \code{cdb$count <- N}
since version 0.6.  The function writes to cdb$res (in opposite to
\code{cdbGetUuid()} whitch writes to \code{cdb$id})
}
\examples{
\dontrun{
ccc            <- cdbIni()
ccc$count      <- 100
cdbGetUuidS(ccc)$res
}
}
\author{
wactbprot
}
\references{
\url{ http://www.omegahat.org/RCurl/ } \url{
http://www.omegahat.org/RJSONIO/ } \url{ http://couchdb.apache.org/ }
}
\seealso{
\code{cdbMakeDB}
}
\keyword{misc}

