% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdbGetView.R
\name{cdbGetView}
\alias{cdbGetView}
\title{Receive view results from CouchDB}
\usage{
cdbGetView(cdb)
}
\arguments{
\item{cdb}{Beside the connection details (\code{cdb$port},\code{cdb$DAName}
...) the \code{cdb$design} and \code{cdb$view} is needed.}
}
\value{
\item{cdb }{The result of the request is stored in cdb$res after converting
the json answer into a list using fromJSON(). If a needed cdb list entry was
not provided cdb$error says something about the R side

}
}
\description{
The function provides accesses to CouchDB views.
}
\details{
Query params e.g. \code{"reduce=false"} or \code{"group_level=1"} can be
provided in \code{cdb$queryParam}
}
\note{
For the moment only one \code{cdb$queryParam} is possible. In the
future maybe Duncans \code{RJavaScript} package can be used to generate
views without leaving R.
}
\author{
wactbprot
}
\keyword{misc}

