% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdbIni.R
\name{cdbIni}
\alias{cdbIni}
\title{Ini function}
\usage{
cdbIni(serverName="localhost",
port="5984",
prot = "http",
DBName="",
uname = "",
pwd = "",
newDBName = "",
removeDBName = "",
id  = "",
fileName = "",
design = "",
view = "",
list = "",
show = "",
queryParam = "",
encSub = "?",
count = 10,
dataList = list(),
attachmentsWithPath=TRUE,
digits = 7)
}
\arguments{
\item{serverName}{server name}

\item{port}{port}

\item{prot}{name of the protocol default is http}

\item{DBName}{name of database}

\item{uname}{name of the user}

\item{pwd}{password}

\item{newDBName}{name of the database for cdbMakeDB()}

\item{removeDBName}{name of the database to remove with cdbRemoveDB()}

\item{id}{the document id to get, put, post or delete}

\item{fileName}{for use in cdbAddAttachment}

\item{design}{the name of the design used when asking a view or list}

\item{view}{the name of a view to query}

\item{list}{the name of a list to query}

\item{show}{the name of a show to query}

\item{queryParam}{additional query params}

\item{encSub}{a character which is used as a replacement for chars who can not be converted by iconv}

\item{count}{how many uuids should be returned by cdbGetUuidS()}

\item{dataList}{a list containing data to post or update}

\item{attachmentsWithPath}{effects the result of the function cdbAddAttachment in the way the variable is named}

\item{digits}{digits kept at toJSON conversion}
}
\value{
\item{cdb}{The R4CouchDB (method) chain(ing) list }
}
\description{
Function returns a list with some default settings and often used functions
such as \code{cdb$baseUrl}.
}
\details{
The list: \code{ cdb <- list(serverName = "localhost", ... )} is returned if
the packages \code{library(RCurl)} and \code{library(RJSONIO)} are
successfully loaded.
}
\examples{
\dontrun{
ccc <- cdbIni(digits=13,
              DBName="r4couch_db",
              attachmentsWithPath=FALSE,
              dataList=list(normalDistRand =  rnorm(20)))
}
}
\author{
wactbprot, parisni
}
\keyword{misc}

