% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cdbListDB.R
\name{cdbListDB}
\alias{cdbListDB}
\title{Returns all databases on the server}
\usage{
cdbListDB(cdb)
}
\arguments{
\item{cdb}{Only the connection settings \code{cdb$port} and
\code{cdb$serverName} is needed.}
}
\value{
\item{cdb }{The result of the request is stored in cdb$res after converting
the json answer into a list using \code{cdb$fromJSON()}.}
}
\description{
Gives a list of all databases available at \code{cdb$serverName}.
}
\details{
The function uses the \code{_all_dbs}  API end point .
}
\examples{
\dontrun{
cdbListDB(cdbIni())$res
}
}
\author{
wactbprot
}
\seealso{
\code{cdbMakeDB}
}
\keyword{misc}

